+++ tests/repo-test.js
@@ -6,7 +6,7 @@ var should = chai.should;
 
 var repo = require(process.cwd()+'/repo');
 var stubs = require('./stubs.js');
-var db = stubs.db();
+var db = stubs.db;
 
 // create a testData.js and put it in /private
 // take a few users from the db and add them in an object replacing their
@@ -18,20 +18,34 @@ describe('Firebase acccess tests', function(){
   var testUser1  = testData.user1;
 
   before(function(done){
-    var updateRef = db.ref('test/users').child(testUser1.id);
-    updateRef.update(testUser1).then(function(){
-      var killUser = db.ref('test/users').child("111000");
-      killUser.set(null, function(err){
+    var eraseUser = db.ref('dev/users').child(testUser1.id);
+    eraseUser.remove()
+      .then(function(){
         done();
       });
-    }).catch(function(err){
-      console.log(err);
+  });
+  
+  it('Should log user as a new user', function(done){
+    repo.logUser(db, testUser1, function(user){
+      console.log(user);
+      expect(user.logType).to.equal("inserted");
+      expect(user.props).to.equal(0);
       done();
     });
   });
-  
+
+  it('Should log user as a updated user', function(done){
+    testUser1.dubs = 51000;
+    repo.logUser(db, testUser1, function(user){
+      expect(user.logType).to.equal("updated");
+      expect(user.dubs).to.equal(51000);
+      done();
+    });
+  });
+
   it('Should add one prop point to user', function(done){
     repo.incrementUser(db, testUser1, "props", function(user){
+      expect(user).to.be.a("object");
       expect(user.props).to.equal(1);
       done();
     });
@@ -39,6 +53,7 @@ describe('Firebase acccess tests', function(){
 
   it('Should add one flow point to user', function(done){
     repo.incrementUser(db, testUser1, "flow", function(user){
+      expect(user).to.be.a("object");
       expect(user.flow).to.equal(1);
       done();
     });
@@ -46,64 +61,69 @@ describe('Firebase acccess tests', function(){
 
   it('Should find user by ID', function(done){
     repo.findUserById(db, testUser1.id, function(user){
-      expect(user.flow).to.equal(1);
       expect(user.username).to.equal(testUser1.username);
       expect(user).to.be.a("object");
       done();
     });
   });
+});
 
-  it('Should update user info', function(done){
-      var newdata = {
-        'dubs': 100,
-        'flow' : 300,
-        'props' : 2
-      };
-    repo.updateUser(db, testUser1.id, newdata, function(result){
-      repo.findUserById(db, testUser1.id, function(user){
-        expect(user.flow).to.equal(300);
-        expect(user.dubs).to.equal(100);
-        expect(user.props).to.equal(2);
-        done();
-      });
-    });
-  });
+describe("Firebase Trigger test", function(){
 
-  it('Should insert new user data', function(done){
-    var newUser = {
-      'dubs': 500,
-      'id' : "111000",
-      "username": "stupidface"
-    };
-    repo.insertUser(db, newUser, function(result){
-      repo.findUserById(db, "111000", function(user){
-        expect(user.flow).to.equal(0);
-        expect(user.dubs).to.equal(500);
-        expect(user.props).to.equal(0);
-        expect(user.username).to.equal("stupidface");
-        done();
-      });
-    });
-  });
-
-  it('Should get top 3 flow leaders of points', function(done){
-    repo.getLeaders(db, "flow", 3, function(items){
-      expect(items).to.be.a('object');
-      expect(items[testUser1.id].flow).to.equal(300);
+  var testTriggerKey = null;
+  it('Should get trigger by triggerName', function(done){
+    repo.getTrigger({}, db, 'test', function(snapshot){
+      var val = snapshot.val();
+      var keys = Object.keys(val);
+      testTriggerKey = keys[0];
+      var theReturn = val[testTriggerKey].Returns;
+      expect(theReturn).to.equal('this is a test 2');
       done();
     });
   });
 
-  // it('Should get trigger by triggerName', function(done){});
+  it('Should update trigger by triggerName', function(done){
+    var data = {};
+    data.triggerName = 'test';
+    data.triggerText = 'I am the new thing';
+    data.user = { username: 'person' };
 
-  // it('Should update trigger by triggerName', function(done){});
+    repo.updateTrigger(db, data, testTriggerKey)
+      .then(function(snapshot){
+        var val = snapshot.val();
+        console.log('update', val);
+        var keys = Object.keys(val);
+        var theReturn = val[keys[0]].Returns;
+        expect(theReturn).to.equal('I am the new thing');
+        done();
+    });
+  });
 
-  // it('Should create trigger by triggerName', function(done){});
+  it('Should delete trigger by trigger key', function(done){
+    repo.deleteTrigger(db, testTriggerKey)
+      .then(function(){
+        db.ref('triggers/' + testTriggerKey).once('value', function(snap){
+          should.not.exist(snap.val());
+          done();
+        });
+      });
+  });
 
-  // it('Should delete trigger by triggerName', function(done){});
+  it('Should create trigger by triggerName', function(done){
+    var data = {};
+    data.triggerName = 'test';
+    data.triggerText = 'this is a test 2';
+    data.user = { username: 'person' };
+    repo.insertTrigger(db, data)
+      .then(function(){
+        repo.getTrigger({}, db, 'blabla', function(val){
+            var keys = Object.keys(val);
+            var theReturn = val[keys[0]].Returns;
+            expect(theReturn).to.equal('this is a test 2');
+            done();
+          });
+      });
+  });
 
-  // it('Should insert a song issue', function(done){
-
-  // });
-
+  
 });